Improved Layer Colour Blending
Layering height maps now produces nicer colour blending between layers, this is done by getting the difference between the current output heightmap and the input map, and clamping this between 0 and 1.
a value of 1 would mean the colour from the input map is used, 0 the current output and somewhere inbetween a blend of the two colours.