Written a mesh job for the ABVC Shader
Written mesh job to encode all the required data for the advanced blending vertex colour shader.
This job writes to 4 uv maps with float4s for all, additionally it writes to the colour map with 3 values. This uses 3 data stream, the UVs being written to on stream 1 with pos and colour the other 2.
Add the extra data fileds to the HeightMapElement struct.